+++ types/contracts/dao/query_msg.ts
@@ -1,75 +1,16 @@
-import {Addr, Denom, Status} from "./shared-types";
+import {Addr, Denom, Proposal_1, Status, Vote_1, Votes_1,} from "./shared-types";
 
 export type QueryMsg =
-  | {
-      get_config: {
-        [k: string]: unknown;
-      };
-    }
-  | {
-      token_list: {
-        [k: string]: unknown;
-      };
-    }
-  | {
-      token_balances: {
-        limit?: number | null;
-        order?: RangeOrder | null;
-        start?: Denom | null;
-        [k: string]: unknown;
-      };
-    }
-  | {
-      proposal: {
-        proposal_id: number;
-        [k: string]: unknown;
-      };
-    }
-  | {
-      proposals: {
-        limit?: number | null;
-        order?: RangeOrder | null;
-        query: ProposalsQueryOption;
-        start?: number | null;
-        [k: string]: unknown;
-      };
-    }
-  | {
-      proposal_count: {
-        [k: string]: unknown;
-      };
-    }
-  | {
-      vote: {
-        proposal_id: number;
-        voter: string;
-        [k: string]: unknown;
-      };
-    }
-  | {
-      votes: {
-        limit?: number | null;
-        order?: RangeOrder | null;
-        proposal_id: number;
-        start?: string | null;
-        [k: string]: unknown;
-      };
-    }
-  | {
-      deposit: {
-        depositor: string;
-        proposal_id: number;
-        [k: string]: unknown;
-      };
-    }
-  | {
-      deposits: {
-        limit?: number | null;
-        order?: RangeOrder | null;
-        query: DepositsQueryOption;
-        [k: string]: unknown;
-      };
-    };
+  | GetConfig
+  | TokenList
+  | TokenBalances
+  | Proposal_1
+  | Proposals
+  | ProposalCount
+  | Vote_1
+  | Votes_1
+  | Deposit
+  | Deposits;
 export type RangeOrder = "asc" | "desc";
 export type ProposalsQueryOption =
   | {
@@ -110,3 +51,102 @@ export type DepositsQueryOption =
         [k: string]: unknown;
       };
     };
+
+/**
+ * Returns [ConfigResponse]
+ *
+ * ## Example
+ *
+ * ```json { "get_config": {} } ```
+ */
+export interface GetConfig {
+  get_config: {
+    [k: string]: unknown;
+  };
+}
+/**
+ * Queries list of cw20 Tokens associated with the DAO Treasury. Returns [TokenListResponse]
+ *
+ * ## Example
+ *
+ * ```json { "token_list": {} } ```
+ */
+export interface TokenList {
+  token_list: {
+    [k: string]: unknown;
+  };
+}
+/**
+ * Returns [TokenBalancesResponse] All DAO Cw20 Balances
+ *
+ * ## Example
+ *
+ * ```json { "token_balances": { "start"?: { "native": "uosmo" | "cw20": "osmo1deadbeef" }, "limit": 30 | 10, "order": "asc" | "desc" } } ```
+ */
+export interface TokenBalances {
+  token_balances: {
+    limit?: number | null;
+    order?: RangeOrder | null;
+    start?: Denom | null;
+    [k: string]: unknown;
+  };
+}
+
+/**
+ * Returns [ProposalsResponse]
+ *
+ * ## Example
+ *
+ * ```json { "proposals": { "query": { "find_by_status": { "status": "pending" | .. | "executed" } | "find_by_proposer": { "proposer": "osmo1deadbeef" } | "everything": {} }, "start"?: 10, "limit": 30 | 10, "order": "asc" | "desc" } } ```
+ */
+export interface Proposals {
+  proposals: {
+    limit?: number | null;
+    order?: RangeOrder | null;
+    query: ProposalsQueryOption;
+    start?: number | null;
+    [k: string]: unknown;
+  };
+}
+/**
+ * Returns the number of proposals in the DAO (u64)
+ *
+ * ## Example
+ *
+ * ```json { "proposal_count": {} } ```
+ */
+export interface ProposalCount {
+  proposal_count: {
+    [k: string]: unknown;
+  };
+}
+
+/**
+ * Queries single deposit info by proposal id & address of depositor. Returns [DepositResponse]
+ *
+ * ## Example
+ *
+ * ```json { "deposit": { "proposal_id": 1, "depositor": "osmo1deadbeef" } } ```
+ */
+export interface Deposit {
+  deposit: {
+    depositor: string;
+    proposal_id: number;
+    [k: string]: unknown;
+  };
+}
+/**
+ * Queries multiple deposits info by 1. proposal id 2. depositor address Returns [DepositsResponse]
+ *
+ * ## Example
+ *
+ * ```json { "deposits": { "query": { "find_by_proposal": { "proposal_id": 1, "start"?: "osmo1deadbeef" } | "find_by_depositor": { "depositor": "osmo1deadbeef", "start"?: 1 } | "everything": { "start"?: [1, "osmo1deadbeef"] } }, "limit": 30 | 10, "order": "asc" | "desc" } } ```
+ */
+export interface Deposits {
+  deposits: {
+    limit?: number | null;
+    order?: RangeOrder | null;
+    query: DepositsQueryOption;
+    [k: string]: unknown;
+  };
+}
